integrations/telemetry/ops/mlops_pipeline.py
"""
mlops_pipeline.py
Pipeline MLOps completo para treinamento de AGI
"""

import mlflow
try:
    import wandb
    from kubeflow import pipelines
except ImportError:
    wandb = None
    pipelines = None
import asyncio
from typing import List, Dict

class AGI_MLOpsPipeline:
    """Pipeline MLOps para treinamento contínuo de AGI"""

    def __init__(self):
        self.mlflow_tracking = "http://mlflow.crux86.svc.cluster.local:5000"
        self.wandb_project = "project-crux86"

        # Componentes do pipeline
        self.components = {
            'data_collection': self._data_collection_component,
            'preprocessing': self._preprocessing_component,
            'model_training': self._training_component,
            'evaluation': self._evaluation_component,
            'safety_check': self._safety_component,
            'deployment': self._deployment_component,
        }

    def _data_collection_component(self, telemetry_sources, hours_to_collect=24): return "/data/telemetry/raw"
    def _preprocessing_component(self, raw_data_dir, output_dir): return output_dir
    def _training_component(self, train_data_dir, val_data_dir, model_config): return "/models/final"
    def _evaluation_component(self, model_path): return {}
    def _safety_component(self, model_path): return {}
    def _deployment_component(self, model_path): return True

    async def _parallel_collect(self, collectors): pass
    def _validate_telemetry_integrity(self, dir): pass

class TelemetryCollector:
    def __init__(self, **kwargs): pass

class TelemetryProcessor:
    def __init__(self, **kwargs): pass
    def extract_features(self): return []
    def normalize(self, f): return f
    def create_sequences(self, n, **kwargs): return []
    def split_dataset(self, s): return {}
    def compute_statistics(self, s): return {}

class ModelSafetyChecker:
    def __init__(self, path): pass
    def check_alignment(self, **kwargs): return 1.0
    def test_goal_inversion(self): return True
    def test_robustness(self, **kwargs): return 1.0
    def monitor_entropy(self): return {}
    def run_byzantine_fire_test(self): return {}
    def generate_safety_report(self): return ""
    def passes_all_tests(self): return True

class PipelineFailedError(Exception): pass
